Update docs and examples
