+++ README.md
@@ -2,7 +2,7 @@
 
 Logger4s is a wrapping [SLF4J](https://www.slf4j.org/) purely functional library for Scala. 
 It's easy to use and does not force a specific target context. 
-You can run your computations in any type `F[_]` that has an instance of cats-effect's `Sync[F]`.
+You can run your computations in any type `F[_]`.
 
 ## Prerequisites ##
 
@@ -16,41 +16,56 @@ Logger4s is published to Sonatype OSS and Maven Central:
 
 - Group id / organization: *org.pure4s*
 - Artifact id / name: *logger4s*
-- Latest version is 0.2.0
+- Latest version is 0.3.0
 
 Usage with SBT, adding a dependency to the latest version of Logger4s to your `build.sbt`:
 
 ```scala
 // For Scala 2.11, or 2.12
-libraryDependencies += "org.pure4s" %% "logger4s" % "0.2.0"
+libraryDependencies += "org.pure4s" %% "logger4s-core"   % "0.3.0"  // Only if you want to support any backend
+libraryDependencies += "org.pure4s" %% "logger4s-cats"   % "0.3.0"  // Cats ecosystem (cats-effect)
+libraryDependencies += "org.pure4s" %% "logger4s-scalaz" % "0.3.0"  // Scalaz ecosystem (scalaz-zio)
 ```
 
 ## Using Logger4s ##
 
-Basic example:
+Example with LazyLogging and Cats:
+```scala
+import cats.effect.IO
+import org.pure4s.logger4s.LazyLogging
+import org.pure4s.logger4s.cats.Logger
+import org.pure4s.logger4s.cats.Logger._
+
+object Main extends App with LazyLogging {
+  Logger[IO].info(s"Hello word, functional logger").unsafeRunSync()
+  //2019-03-03 21:34:04.880 [BasicExampleMain$][INFO ] Hello word, purely functional logger
+}
+```
+
+Example without LazyLogging and Cats:
 ```scala
 import cats.effect.{IO, Sync}
 import cats.implicits._
-import org.pure4s.logger4s.Logger
+import org.pure4s.logger4s.cats.Logger
 
 case class User(email: String)
 
 class UserService[F[_] : Sync : Logger] {
   def findByEmail(email: String): F[Option[User]] = {
-    Logger[F].info(s"Hello word, functional logger ($email)") *> Option(User(email)).pure[F]
+    Logger[F].info(s"User email is $email") *> Option(User(email)).pure[F]
   }
 }
 
-object BasicExampleMain extends App {
+object Main extends App {
   implicit val instance: Logger[IO] = Logger.instance[IO](classOf[UserService[IO]])
 
   val service = new UserService[IO]
   service.findByEmail("example@example.com").unsafeRunSync()
-  //2019-01-27 21:40:40.557 [UserService][INFO] - Hello word, functional logger (example@example.com)
+  //2019-03-03 21:35:35.286 [UserService][INFO ] User email is example@example.com 
 }
 ```
 
-Basic example with LazyLogging
+Example for comprehensions with LazyLogging and Cats
 ```scala
 import cats.effect.{IO, Sync}
 import cats.implicits._
@@ -78,7 +93,7 @@ class AuthService[F[_] : Sync] extends LazyLogging {
   }
 }
 
-object BasicLazyLoggingExampleMain extends App {
+object Main extends App {
   val service = new AuthService[IO]
   service.login("example@example.com","123").unsafeRunSync()
   //2019-01-27 21:40:40.557 [AuthService][INFO] - Login with email = example@example.com and password = 123
@@ -86,7 +101,7 @@ object BasicLazyLoggingExampleMain extends App {
 }
 ```
 
-Complex example:
+Advance example, custom `Show` with LazyLogging and Cats:
 ```scala
 import cats.Show
 import cats.effect.{IO, Sync}
@@ -115,10 +130,23 @@ class ClientService[F[_] : Sync] extends LazyLogging{
   }
 }
 
-object ComplexExampleMain extends App {
+object Main extends App {
   val service = new ClientService[IO]
   service.findByEmail("example@example.com").unsafeRunSync()
   //2019-01-27 21:25:26.150 [ClientService][INFO] - {"email":"example@example.com"}
+}
+```
+
+Example with LazyLogging and Scalaz:
+```scala
+import org.pure4s.logger4s.LazyLogging
+import scalaz.zio.{IO, RTS}
+import org.pure4s.logger4s.scalaz.Logger
+
+object Main extends App with RTS with LazyLogging {
+
+  unsafeRun(Logger[IO[Nothing, ?]].info(s"Hello word, purely functional logger"))
+  //2019-03-03 21:49:59.905 [BasicExampleMain$][INFO ] Hello word, purely functional logger
 }
 ```
