+++ example-cats/src/main/scala/BasicExampleMain.scala
@@ -1,18 +1,8 @@
-import cats.effect.{IO, Sync}
-import cats.implicits._
+import cats.effect.IO
+import org.pure4s.logger4s.LazyLogging
 import org.pure4s.logger4s.cats.Logger
+import org.pure4s.logger4s.cats.Logger._
 
-case class User(email: String)
-
-class UserService[F[_] : Sync : Logger] {
-  def findByEmail(email: String): F[Option[User]] = {
-    Logger[F].info(s"Hello word, functional logger ($email)") *> Option(User(email)).pure[F]
-  }
-}
-
-object BasicExampleMain extends App {
-  implicit val instance: Logger[IO] = Logger.instance[IO](classOf[UserService[IO]])
-
-  val service = new UserService[IO]
-  service.findByEmail("example@example.com").unsafeRunSync()
+object BasicExampleMain extends App with LazyLogging {
+  Logger[IO].info(s"Hello word, purely functional logger").unsafeRunSync()
 }
